requirements.txt
fastapi
uvicorn
langchain
openai
pydantic
python-dotenv
langchain-community
langchain-openai
axios
tailwindcss postcss autoprefixer
tailwindcss init -p
vite